refactor[task3]: split handler
