+++ product-service/handlers/getProductById/index.ts
@@ -1,0 +1,1 @@
+export { getProduct } from './handler';
