+++ product-service/handlers/getProductsList/index.ts
@@ -1,0 +1,1 @@
+export { getProducts } from './handler';
